+++ python/bosdyn-mission/src/bosdyn/mission/constants.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2019 Boston Dynamics, Inc.  All rights reserved.
+#
+# Downloading, reproducing, distributing or otherwise using the SDK Software
+# is subject to the terms and conditions of the Boston Dynamics Software
+# Development Kit License (20191101-BDSDK-SL).
+
+"""Constants relevant to the missions service."""
+
+SUCCESS_ANSWER_CODE = 0
+FAILURE_ANSWER_CODE = -1
+valid_answer_codes = set((SUCCESS_ANSWER_CODE, FAILURE_ANSWER_CODE))